resumitory-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
supabase==2.0.0
pyjwt==2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
